Create Mini Calendar
In this JavaScript project we will make a mini calendar with the help of HTML, CSS and JavaScript. This calendar will display the today's date, day, month and year. This date will be updated every day. In this project we will use the "Date Object" of JavaScript to get the today's date and display on our website.